refactor(Lotto) : 로또 객체 패키징 변경
