+++ src/test/java/lotto/domain/model/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
